+++ src/components/cinemas/SearchBar.scss
@@ -1,0 +1,7 @@
+.search-bar {
+  display: flex;
+  padding: 0.25rem;
+  background-color: var(--contrast-dark);
+  justify-content: space-around;
+  align-items: center;
+}
